Deconstruct the monolithic GroupLifecyclePage into separate, focused admin pages with specific URLs to improve navigation, maintainability, and user experience. committed
