+++ frontend/src/components/ui/index.js
@@ -1,0 +1,3 @@
+// Barrel export for UI components
+export { SectionCard } from './SectionCard'
+export { StatusBlock } from './StatusBlock'
