+++ frontend/src/pages/admin/MembersPage.jsx
@@ -1,30 +1,10 @@
 import { useEffect, useState } from 'react'
 import apiClient from '../../utils/apiClient'
 import apiConfig from '../../config/api'
+import EntitySearchSelect from '../../components/EntitySearchSelect'
+import { SectionCard, StatusBlock } from '../../components/ui'
 import styles from '../GroupLifecyclePage.module.css'
 import { useAdminGroup } from '../../context/AdminGroupContext'
-
-function SectionCard({ title, description, children }) {
-  return (
-    <div className={styles.sectionCard}>
-      <div className={styles.sectionHeader}>
-        <h2>{title}</h2>
-        <p>{description}</p>
-      </div>
-      <div className={styles.sectionBody}>{children}</div>
-    </div>
-  )
-}
-
-function StatusBlock({ title, message, type }) {
-  if (!message) return null
-  return (
-    <div className={`${styles.statusBlock} ${type === 'error' ? styles.error : styles.success}`}>
-      <strong>{title}</strong>
-      <span>{message}</span>
-    </div>
-  )
-}
 
 function MembersPage() {
   const { currentGroupId } = useAdminGroup()
@@ -71,15 +51,17 @@ function MembersPage() {
               required
             />
           </label>
-          <label>
-            Member User ID
-            <input
-              value={memberUserId}
-              onChange={(e) => setMemberUserId(e.target.value)}
-              placeholder="UUID of member"
-              required
-            />
-          </label>
+          <EntitySearchSelect
+            label="Member"
+            endpoint={apiConfig.endpoints.userSearch}
+            searchField="email"
+            returnField="_id"
+            displayField="email"
+            value={memberUserId}
+            onChange={setMemberUserId}
+            placeholder="Search user by email"
+            required
+          />
           <button type="submit" disabled={memberStatus.loading}>
             {memberStatus.loading ? 'Adding…' : 'Add Member'}
           </button>
